change sabotage to duel
